pyproject.toml
[tool.poetry]
name = "pocket-api"
version = "0.1.0"
description = "api for task manager"
authors = ["rubbie kelvin voltsman <rubbiekelvinvoltsman@gmail.com>"]
license = "GNU"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1"
django-rest-framework = "^0.1.0"

[tool.poetry.dev-dependencies]
ipython = "^8.4.0"
requests = "^2.28.1"
black = "^22.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
